# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Philips Hue Control API
  version: 2.0.0
  description: |2-

          A simplified, client-friendly API for controlling Philips Hue lights.

          **Features:**
          - Pre-computed RGB colors and CSS shadows
          - Unified dashboard endpoint (1 call instead of 4-6)
          - Room hierarchy already built
          - Session-based authentication
          - Demo mode for testing without a real bridge
          - Settings and weather API
          - Real-time updates via WebSocket

          **Authentication Methods:**
          1. **Session Token**: Get a session token from `POST /api/v1/auth/session`, then use `Authorization: Bearer <token>` header
          2. **Demo Mode**: Add `X-Demo-Mode: true` header to use mock data (no bridge required)

          **Demo Mode:**
          Try the API without a Hue Bridge by adding `X-Demo-Mode: true` header to any request.
          - Returns mock data (2 rooms, 6 lights, 4 scenes, 2 zones)
          - State changes persist in memory (reset on server restart)
          - WebSocket: authenticate with `{ type: 'auth', demoMode: true }`

          **Multi-Client Support:**
          When one client pairs with a bridge, credentials are stored server-side. Other clients can then connect without re-pairing:
          1. First client: `POST /api/v1/auth/pair` → `POST /api/v1/auth/session`
          2. Subsequent clients: `POST /api/v1/auth/connect` (no pairing needed!)
...

...

  contact:
...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Session Token
      description: Session token from POST /api/v1/auth/session
    DemoMode:
      type: apiKey
      in: header
      name: X-Demo-Mode
      description: Set to "true" to use demo mode (no bridge required)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: resource_not_found
        message:
          type: string
          example: The light 'abc123' was not found
        suggestion:
          type: string
          example: Check the light ID or refresh the dashboard to get current lights
...
paths:
  /auth/pair:
    post:
      summary: Pair with a Hue Bridge
      description: Initial pairing with a Hue Bridge. User must press the link button on the bridge before calling this endpoint.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bridgeIp
              properties:
                bridgeIp:
                  type: string
                  example: 192.168.1.100
                  description: Your Hue Bridge IP address
                appName:
                  type: string
                  example: hue_control_app
                  description: Application name for the Hue Bridge (optional)
      responses:
        '200':
          description: Pairing successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: KZLwQGOGUMwtFMpYMeRqA9gzSuYoRtRgVThntanx
                    description: Generated Hue API key
        '400':
          description: Link button not pressed or invalid bridge IP
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: link button not pressed
                  type:
                    type: integer
                    example: 101
        '504':
          description: Bridge connection timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/connect:
    post:
      summary: Connect using stored credentials
      description: Connect to a bridge using server-side stored credentials. Use this when another client has already paired with the bridge. No pairing needed!
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bridgeIp
              properties:
                bridgeIp:
                  type: string
                  example: 192.168.1.100
                  description: Your Hue Bridge IP address
      responses:
        '200':
          description: Connected successfully using stored credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionToken:
                    type: string
                    example: hue_sess_abc123...
                  expiresIn:
                    type: integer
                    example: 86400
                  bridgeIp:
                    type: string
                    example: 192.168.1.100
        '401':
          description: Stored credentials are no longer valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Stored credentials are no longer valid. Pairing required.
                  requiresPairing:
                    type: boolean
                    example: true
        '404':
          description: No stored credentials for this bridge
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No stored credentials for this bridge. Pairing required.
                  requiresPairing:
                    type: boolean
                    example: true
  /auth/bridge-status:
    get:
      summary: Check if bridge has stored credentials
      description: Check whether the server has stored credentials for a specific bridge IP
      tags:
        - Authentication
      parameters:
        - name: bridgeIp
          in: query
          required: true
          schema:
            type: string
          description: Bridge IP to check
          example: 192.168.1.100
      responses:
        '200':
          description: Bridge status
          content:
            application/json:
              schema:
                type: object
                properties:
                  bridgeIp:
                    type: string
                    example: 192.168.1.100
                  hasCredentials:
                    type: boolean
                    example: true
        '400':
          description: Missing bridgeIp parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/session:
    post:
      summary: Create a new session
      description: Connect to your Hue Bridge and get a session token for subsequent requests
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bridgeIp
                - username
              properties:
                bridgeIp:
                  type: string
                  example: 192.168.1.100
                  description: Your Hue Bridge IP address
                username:
                  type: string
                  example: abc123def456
                  description: Your Hue API username/key
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionToken:
                    type: string
                    example: hue_sess_abc123...
                  expiresIn:
                    type: integer
                    example: 86400
                    description: Session expires in seconds (24 hours)
                  bridgeIp:
                    type: string
                    example: 192.168.1.100
        '400':
          description: Missing credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Cannot connect to bridge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get current session info
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Session info
          content:
            application/json:
              schema:
                type: object
                properties:
                  bridgeIp:
                    type: string
                    example: 192.168.1.100
                  active:
                    type: boolean
                    example: true
        '401':
          description: Invalid or expired session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Revoke session (logout)
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Session revoked
  /auth/refresh:
    post:
      summary: Refresh session token
      description: Extend the expiration of your current session by getting a new token
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Session refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionToken:
                    type: string
                    example: hue_sess_new123...
                  expiresIn:
                    type: integer
                    example: 86400
                  bridgeIp:
                    type: string
                    example: 192.168.1.100
        '401':
          description: Invalid or expired session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...